Fix persona stats import and organize persona components
