+++ src/features/catalog/personas/components/editor/PersonaColorsTab.tsx
@@ -1,10 +1,10 @@
 import { useState } from "react";
 import { Loader2, Palette, User } from "lucide-react";
 import { toast } from "sonner";
-import { extractColorsFromImage } from "../../../../shared/lib/avatar-color-extraction";
-import { ColorPicker } from "../../../../shared/components/ui/ColorPicker";
-import { cn } from "../../../../shared/lib/utils";
-import type { PersonaFormData } from "../lib/persona-editor-model";
+import { extractColorsFromImage } from "../../../../../shared/lib/avatar-color-extraction";
+import { ColorPicker } from "../../../../../shared/components/ui/ColorPicker";
+import { cn } from "../../../../../shared/lib/utils";
+import type { PersonaFormData } from "../../lib/persona-editor-model";
 import { PersonaEditorSectionHeader } from "./PersonaEditorSectionHeader";
 
 export function PersonaColorsTab({
