+++ src/features/catalog/personas/components/modals/CreatePersonaModal.tsx
@@ -2,9 +2,9 @@
 // Modal: Create Persona (name only)
 // ──────────────────────────────────────────────
 import { useState } from "react";
-import { Modal } from "../../../../shared/components/ui/Modal";
-import { useCreatePersona } from "../hooks/use-personas";
-import { useUIStore } from "../../../../shared/stores/ui.store";
+import { Modal } from "../../../../../shared/components/ui/Modal";
+import { useCreatePersona } from "../../hooks/use-personas";
+import { useUIStore } from "../../../../../shared/stores/ui.store";
 import { Loader2, User } from "lucide-react";
 import { toast } from "sonner";
