+++ src/features/catalog/personas/shell.ts
@@ -1,5 +1,5 @@
-export * from "./components/CreatePersonaModal";
-export * from "./components/ImportPersonaModal";
-export * from "./components/PersonaEditor";
-export * from "./components/PersonaMakerModal";
-export * from "./components/PersonasPanel";
+export * from "./components/modals/CreatePersonaModal";
+export * from "./components/modals/ImportPersonaModal";
+export * from "./components/editor/PersonaEditor";
+export * from "./components/modals/PersonaMakerModal";
+export * from "./components/panel/PersonasPanel";
